+++ src/antibot/antibot.h
@@ -17,7 +17,7 @@ struct UserStateItem {
 	std::string address;
 	int64_t user_registration_date;
 	int64_t address_registration_date;
-	int reputation;
+	double reputation;
 	int64_t balance;
 	bool trial;
 
@@ -152,7 +152,7 @@ struct BlockVTX {
 class AntiBot
 {
 private:
-	void getMode(std::string _address, ABMODE &mode, int &reputation, int64_t &balance, int height);
+	void getMode(std::string _address, ABMODE &mode, double &reputation, int64_t &balance, int height);
 	void getMode(std::string _address, ABMODE &mode, int height);
 	int getLimit(CHECKTYPE _type, ABMODE _mode, int height);
